getNumberOfInvariantFactors -> countInvariantFactors.

+++ engine/algebra/nabeliangroup.cpp
@@ -37,8 +37,7 @@
 
 namespace regina {
 
-const NLargeInteger& NAbelianGroup::getInvariantFactor(
-        unsigned long index) const {
+const NLargeInteger& NAbelianGroup::getInvariantFactor(size_t index) const {
     std::multiset<NLargeInteger>::const_iterator it = invariantFactors.begin();
     advance(it, index);
     return *it;
